+++ Web/SFB.Web.UI/Controllers/BenchmarkCriteriaController.cs
@@ -53,7 +53,7 @@ namespace SFB.Web.UI.Controllers
         /// </summary>
         /// <param name="urn"></param>
         /// <returns></returns>
-        public async Task<ViewResult> ComparisonStrategy(int urn)
+        public async Task<ViewResult> ComparisonStrategy(long urn)
         {
             ViewBag.URN = urn;
 
@@ -66,7 +66,7 @@ namespace SFB.Web.UI.Controllers
             return View(benchmarkSchool);
         }
 
-        private async Task<ActionResult> ErrorViewForComparisonStrategy(int urn)
+        private async Task<ActionResult> ErrorViewForComparisonStrategy(long urn)
         {
             ViewBag.URN = urn;
 
@@ -82,7 +82,7 @@ namespace SFB.Web.UI.Controllers
         
         }
 
-        public async Task<ActionResult> StepOne(int? urn, int?  fuid, ComparisonType? comparisonType, bool similarPupils = false)
+        public async Task<ActionResult> StepOne(long? urn, long? fuid, ComparisonType? comparisonType, bool similarPupils = false)
         {
             switch (comparisonType)
             {
@@ -116,7 +116,7 @@ namespace SFB.Web.UI.Controllers
             }
         }
 
-        private async Task<ViewResult> HowWeCalculateSpecials(int urn, bool similarPupils)
+        private async Task<ViewResult> HowWeCalculateSpecials(long urn, bool similarPupils)
         {
             ViewBag.URN = urn;
             ViewBag.SimilarPupils = similarPupils;
@@ -132,7 +132,7 @@ namespace SFB.Web.UI.Controllers
         /// <param name="urn"></param>
         /// <param name="comparisonType"></param>
         /// <returns></returns>
-        public async Task<ViewResult> SelectBasketSize(int urn, ComparisonType comparisonType)
+        public async Task<ViewResult> SelectBasketSize(long urn, ComparisonType comparisonType)
         {
             ViewBag.URN = urn;
             ViewBag.ComparisonType = comparisonType;
@@ -149,7 +149,7 @@ namespace SFB.Web.UI.Controllers
         /// <param name="comparisonType"></param>
         /// <param name="estType"></param>
         /// <returns></returns>
-        public async Task<ViewResult> SelectSchoolType(int? urn, int? fuid, ComparisonType comparisonType, EstablishmentType? estType, int? basketSize)
+        public async Task<ViewResult> SelectSchoolType(long? urn, long? fuid, ComparisonType comparisonType, EstablishmentType? estType, int? basketSize)
         {
             ViewBag.URN = urn;
             ViewBag.Fuid = fuid;
@@ -183,7 +183,7 @@ namespace SFB.Web.UI.Controllers
         /// <param name="comparisonType"></param>
         /// <param name="estType"></param>
         /// <returns></returns>
-        public async Task<ViewResult> ChooseRegion(int? urn, ComparisonType comparisonType, EstablishmentType? estType, bool excludePartial = false)
+        public async Task<ViewResult> ChooseRegion(long? urn, ComparisonType comparisonType, EstablishmentType? estType, bool excludePartial = false)
         {
             ViewBag.URN = urn;
             ViewBag.ComparisonType = comparisonType;
@@ -236,7 +236,7 @@ namespace SFB.Web.UI.Controllers
         /// <param name="laCodeName"></param>
         /// <returns></returns>
         public async Task<ActionResult> AdvancedCharacteristics(
-            int? urn, 
+            long? urn, 
             ComparisonType comparisonType, 
             EstablishmentType estType, 
             ComparisonArea? areaType, 
@@ -350,7 +350,7 @@ namespace SFB.Web.UI.Controllers
         /// </summary>
         /// <param name="urn"></param>
         /// <returns></returns>
-        public ActionResult HighestProgressSchoolsBenchmarking(int urn)
+        public ActionResult HighestProgressSchoolsBenchmarking(long urn)
         {
             ViewBag.URN = urn;
             return View("HighestProgressSchoolsBenchmarking");
@@ -362,7 +362,7 @@ namespace SFB.Web.UI.Controllers
         /// <param name="urn"></param>
         /// <returns></returns>
         [ValidateAntiForgeryToken]
-        public async Task<ActionResult> BestInClassCharacteristics(int urn, BestInClassCriteria bicCriteria)
+        public async Task<ActionResult> BestInClassCharacteristics(long urn, BestInClassCriteria bicCriteria)
         {                                
             var benchmarkSchool = await InstantiateBenchmarkSchoolAsync(urn);
 
@@ -380,7 +380,7 @@ namespace SFB.Web.UI.Controllers
         /// <param name="estType"></param>
         /// <param name="simpleCriteria"></param>
         /// <returns></returns>
-        public async Task<ActionResult> SimpleCharacteristics(int? urn, int? fuid, ComparisonType comparisonType, EstablishmentType? estType, SimpleCriteria SimpleCriteria)
+        public async Task<ActionResult> SimpleCharacteristics(long? urn, long? fuid, ComparisonType comparisonType, EstablishmentType? estType, SimpleCriteria SimpleCriteria)
         {
             if (estType.HasValue)
             {
@@ -438,7 +438,7 @@ namespace SFB.Web.UI.Controllers
         /// <returns></returns>
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<ActionResult> OverwriteStrategy(int? urn, ComparisonType comparisonType, EstablishmentType estType, BenchmarkCriteriaVM criteria, 
+        public async Task<ActionResult> OverwriteStrategy(long? urn, ComparisonType comparisonType, EstablishmentType estType, BenchmarkCriteriaVM criteria, 
             ComparisonArea areaType, int? lacode, string schoolName, int basketCount, bool excludePartial = false)
         {
             ViewBag.URN = urn;
@@ -516,7 +516,7 @@ namespace SFB.Web.UI.Controllers
             return result;
         }
 
-        private async Task<SchoolViewModel> InstantiateBenchmarkSchoolAsync(int urn)
+        private async Task<SchoolViewModel> InstantiateBenchmarkSchoolAsync(long urn)
         {
             var benchmarkSchool = new SchoolViewModel(await _contextDataService.GetSchoolDataObjectByUrnAsync(urn), _benchmarkBasketService.GetSchoolBenchmarkList());
             var schoolsLatestFinancialDataModel = await _financialDataService.GetSchoolsLatestFinancialDataModelAsync(benchmarkSchool.Id, benchmarkSchool.EstablishmentType);
